+++ CHANGELOG.md
@@ -10,8 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Add setNewRule and unsetNewRule to InlineLexer
-* Add setNewRule and unsetNewRule to BlockLexer
+* Add setRule and unsetRule to InlineLexer and BlockLexer
 * Export: BlockLexer, InlineLexer, Parser
 
 ### Changed